requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.2
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
vaderSentiment>=3.3.2
transformers>=4.36.0
torch>=2.1.0
pyttsx3>=2.90
azure-cognitiveservices-speech>=1.34.0
pyyaml>=6.0.1
aiofiles>=23.2.1
